backend/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
motor
pymongo
numpy
scikit-learn
joblib
